+++ Sources/KeyboardKit/_Deprecated/Image+Deprecated.swift
@@ -2,6 +2,12 @@ import SwiftUI
 
 public extension Image {
     
+    @available(*, deprecated, renamed: "keyboardLeft")
+    static var moveCursorLeft: Image { .keyboardLeft }
+    
+    @available(*, deprecated, renamed: "keyboardRight")
+    static var moveCursorRight: Image { .keyboardRight }
+    
     @available(*, deprecated, renamed: "keyboardSettings")
     static var settings: Image { .keyboardSettings }
 }